Change do_create
Changed to look through command line arguments without shlex first and see if things are supposed to be strings.
If they are not supposed to be a string, they won't contain a " and thus we check for a . to turn into a float. Otherwise turn into an integer.
Then shlex like normal and check through the number_dict when using setattr.